Move load_known_adducts to filter_utils
