+++ matchms/filtering/filter_utils/derive_precursor_mz_and_parent_mass.py
@@ -4,7 +4,8 @@ from matchms.constants import PROTON_MASS
 from matchms.filtering.filter_utils.interpret_unknown_adduct import \
     get_multiplier_and_mass_from_adduct
 from matchms.filtering.metadata_processing.clean_adduct import (
-    _clean_adduct, load_known_adducts)
+    _clean_adduct)
+from matchms.filtering.filter_utils.load_known_adducts import load_known_adducts
 from matchms.typing import SpectrumType
