+++ matchms/filtering/metadata_processing/derive_adduct_from_name.py
@@ -1,8 +1,8 @@
 import logging
 import re
 from matchms.typing import SpectrumType
-from .clean_adduct import _clean_adduct, load_known_adducts
-
+from .clean_adduct import _clean_adduct
+from ..filter_utils.load_known_adducts import load_known_adducts
 
 logger = logging.getLogger("matchms")
