+++ matchms/filtering/metadata_processing/derive_ionmode.py
@@ -1,7 +1,7 @@
 import logging
 from matchms.typing import SpectrumType
-from .clean_adduct import _clean_adduct, load_known_adducts
-
+from .clean_adduct import _clean_adduct
+from ..filter_utils.load_known_adducts import load_known_adducts
 
 logger = logging.getLogger("matchms")
