+++ matchms/filtering/metadata_processing/repair_adduct_based_on_smiles.py
@@ -2,8 +2,7 @@ import logging
 from matchms import Spectrum
 from matchms.filtering.filter_utils.get_neutral_mass_from_smiles import \
     get_monoisotopic_neutral_mass
-from matchms.filtering.metadata_processing.clean_adduct import \
-    load_known_adducts
+from ..filter_utils.load_known_adducts import load_known_adducts
 from .repair_parent_mass_is_mol_wt import repair_parent_mass_is_mol_wt
