+++ tests/filtering/test_filter_utils/test_interpret_unknown_adduct.py
@@ -1,8 +1,7 @@
 import pytest
 from matchms.filtering.filter_utils.interpret_unknown_adduct import \
     get_multiplier_and_mass_from_adduct
-from matchms.filtering.metadata_processing.clean_adduct import \
-    load_known_adducts
+from matchms.filtering.filter_utils.load_known_adducts import load_known_adducts
 
 
 def test_get_multiplier_and_mass():
