+++ tests/filtering/test_load_adducts.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pandas as pd
-from matchms.filtering.metadata_processing.clean_adduct import (
-    load_known_adduct_conversions, load_known_adducts)
+from matchms.filtering.filter_utils.load_known_adducts import load_known_adducts, load_known_adduct_conversions
 
 
 def test_load_adducts_dict():
